Add new components and update UI elements
This commit adds the following new components:
apiService.jsx: This file contains functions that make API requests to the backend microservices.
DataParsing.jsx: This file contains functions for parsing the response data from the microservices.
SubmitReviewForm.jsx: This file contains the UI component with a text field for the user to input a review to be analyzed.
ReviewResult.jsx: This file contains the UI component that displays the result of the sentiment analysis for the user's inputted review.
DropdownComponent.jsx: This file contains the UI component that allows the user to select which dataset they want to analyze.

In addition to the new components, the following UI elements have been updated:
Navigation.jsx: The navigation links on the nav bar have been updated.
Homepage.jsx: The links on the homepage have been updated.
Dashboard.jsx: The newly created DropdownComponent has been integrated into the dashboard.

These changes improve the overall user experience of the application by providing a more user-friendly interface for selecting the desired dataset and analyzing user reviews.